Quick hack for missing LINK
