+++ gcc/d/dmd/root/dsystem.h
@@ -21,4 +21,4 @@
 
 #pragma once
 
-#include "d-system.h"
+#include "d/d-system.h"
